Move code to layers

+++ MoviesDbNode/public/javascripts/page.js
@@ -1,7 +1,6 @@
 
-// Extract updates from page and call the update route
+// Update a movie using numeric index from page
 function updateMovie(id) {
-
     var title = $("#movie-title-" + id).val();
     var format = $("#movie-format-" + id).val();
     var length = +($("#movie-length-" + id).val());
@@ -9,36 +8,40 @@ function updateMovie(id) {
     var rating = +($("#movie-rating-" + id).val());
     var guid = $("#movie-guid-" + id).val();
 
-    var jsonData = {
-        guid: guid,
-        format: format,
-        rating: rating,
-        title: title,
-        releaseYear: releaseYear,
-        length: length
-    };
+    if (guid) {
+        var jsonData = {
+            guid: guid,
+            format: format,
+            rating: rating,
+            title: title,
+            releaseYear: releaseYear,
+            length: length
+        };
 
-    return new Promise(function (resolve, reject) {
-        $.ajax({
-            url: './update',
-            data: JSON.stringify(jsonData),
-            type: "PUT",
-            contentType: 'application/json',
-            dataType: "json"
-        })
-            .done(function (data) {
-                resolve(alert(data));
+        return new Promise(function (resolve, reject) {
+            $.ajax({
+                url: './update',
+                data: JSON.stringify(jsonData),
+                type: "PUT",
+                contentType: 'application/json',
+                dataType: "json"
             })
-            .fail(function (jqXHR) {
-                reject(alert(jqXHR.responseText));
-            })
-            .always(function () {
-                window.location.href = window.location.href;
-            });
-    });
+                .done(function (data) {
+                    resolve(alert(data));
+                })
+                .fail(function (jqXHR) {
+                    reject(alert(jqXHR.responseText));
+                })
+                .always(function () {
+                    window.location.href = window.location.href;
+                });
+        });
+    } else {
+        alert('No GUID found for movie ' + id);
+    }
 }
 
-// Extract updates from page and call the update route
+// Add a movie using numeric index from page
 function addMovie() {
 
     var title = $("#new-movie-title").val();
@@ -75,35 +78,37 @@ function addMovie() {
     });
 }
 
-// Extract updates from page and call the update route
+// Delete a movie using numeric index from page
 function deleteMovie(id) {
-
     var guid = $("#movie-guid-" + id).val();
-
-    return new Promise(function (resolve, reject) {
-        $.ajax({
-            url: './delete/' + guid,
-            type: "DELETE",
-            contentType: 'application/json',
-            dataType: "json"
-        })
-            .done(function (data) {
-                resolve(alert(data));
+    if (guid) {
+        return new Promise(function (resolve, reject) {
+            $.ajax({
+                url: './delete/' + guid,
+                type: "DELETE",
+                contentType: 'application/json',
+                dataType: "json"
             })
-            .fail(function (jqXHR) {
-                reject(alert(jqXHR.responseText));
-            })
-            .always(function () {
-                window.location.href = window.location.href;
-            });
-    });
+                .done(function (data) {
+                    resolve(alert(data));
+                })
+                .fail(function (jqXHR) {
+                    reject(alert(jqXHR.responseText));
+                })
+                .always(function () {
+                    window.location.href = window.location.href;
+                });
+        });
+    } else {
+        alert('No GUID found for movie ' + id);
+    }
 }
 
 // Clear fields on the add table
 function resetAddFields() {
-    var title = $("#new-movie-title").val(null);
-    var format = $("#new-movie-format").val(null);
-    var length = +($("#new-movie-length").val(null));
-    var releaseYear = +($("#new-movie-release-year").val(null));
-    var rating = +($("#new-movie-rating").val(null));
+    $("#new-movie-title").val(null);
+    $("#new-movie-format").val(null);
+    $("#new-movie-length").val(null);
+    $("#new-movie-release-year").val(null);
+    $("#new-movie-rating").val(null);
 }
